backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic>=2
email-validator
pymongo>=4

